Move maths consts to separate util
